Enable glibc-compat to be built with glibc.
Even though the result is an empty package this is needed
not that dependecies are required to be buildable.  Without
this any package that depends on glibc-compat becomes
unbuildable with glibc.

R=binji@chromium.org

Review URL: https://codereview.chromium.org/556373002

+++ ports/glibc-compat/build.sh
@@ -5,11 +5,17 @@
 BUILD_DIR=${SRC_DIR}
 
 ConfigureStep() {
+  if [ "${NACL_LIBC}" != "newlib" ]; then
+    return
+  fi
   LogExecute cp -rf ${START_DIR}/* .
   LogExecute rm -rf out
 }
 
 BuildStep() {
+  if [ "${NACL_LIBC}" != "newlib" ]; then
+    return
+  fi
   # export the nacl tools
   export CC=${NACLCC}
   export CXX=${NACLCXX}
@@ -20,6 +26,9 @@ BuildStep() {
 }
 
 InstallStep() {
+  if [ "${NACL_LIBC}" != "newlib" ]; then
+    return
+  fi
   local LIB=libglibc-compat.a
   INCDIR=${DESTDIR_INCLUDE}/glibc-compat
   MakeDir ${DESTDIR_LIB}
